requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
fastapi==0.66.1
google-auth==1.33.1
h11==0.12.0
idna==3.2
kubernetes==17.17.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
websocket-client==1.1.0
flake8>=3.9.0
flake8-quotes>=3.2.0
